+++ samples/AspNetCoreSample/Program.cs
@@ -6,8 +6,10 @@ using GSS.Authentication.CAS.Validation;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Authentication.OAuth;
+using Microsoft.AspNetCore.Authentication.OpenIdConnect;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http.Extensions;
+using Microsoft.IdentityModel.Protocols.OpenIdConnect;
 using NLog.Web;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -53,6 +55,7 @@ builder.Services.AddAuthentication(CookieAuthenticationDefaults.AuthenticationSc
     .AddCAS(options =>
     {
         options.CasServerUrlBase = builder.Configuration["Authentication:CAS:ServerUrlBase"];
+        options.SaveTokens = builder.Configuration.GetValue("Authentication:CAS:SaveTokens",false);
         var protocolVersion = builder.Configuration.GetValue("Authentication:CAS:ProtocolVersion", 3);
         if (protocolVersion != 3)
         {
@@ -111,6 +114,7 @@ builder.Services.AddAuthentication(CookieAuthenticationDefaults.AuthenticationSc
         options.ClaimActions.MapJsonKey(ClaimTypes.NameIdentifier, "id");
         options.ClaimActions.MapJsonSubKey(ClaimTypes.Name, "attributes", "display_name");
         options.ClaimActions.MapJsonSubKey(ClaimTypes.Email, "attributes", "email");
+        options.SaveTokens = builder.Configuration.GetValue("Authentication:OAuth:SaveTokens",false);
         options.Events = new OAuthEvents
         {
             OnCreatingTicket = async context =>
@@ -135,6 +139,35 @@ builder.Services.AddAuthentication(CookieAuthenticationDefaults.AuthenticationSc
                 using var json = await JsonDocument.ParseAsync(stream).ConfigureAwait(false);
                 context.RunClaimActions(json.RootElement);
             },
+            OnRemoteFailure = context =>
+            {
+                var failure = context.Failure;
+                var logger = context.HttpContext.RequestServices.GetRequiredService<ILogger<OAuthEvents>>();
+                if (!string.IsNullOrWhiteSpace(failure?.Message))
+                {
+                    logger.LogError(failure, "{Exception}", failure.Message);
+                }
+
+                context.Response.Redirect("/Account/ExternalLoginFailure");
+                context.HandleResponse();
+                return Task.CompletedTask;
+            }
+        };
+    })
+    .AddOpenIdConnect(options =>
+    {
+        options.ClientId = builder.Configuration["Authentication:OIDC:ClientId"];
+        options.ClientSecret = builder.Configuration["Authentication:OIDC:ClientSecret"];
+        options.Authority = builder.Configuration["Authentication:OIDC:Authority"];
+        options.MetadataAddress = builder.Configuration["Authentication:OIDC:MetadataAddress"];
+        options.ResponseType = builder.Configuration.GetValue("Authentication:OIDC:ResponseType", OpenIdConnectResponseType.Code);
+        options.ResponseMode = builder.Configuration.GetValue("Authentication:OIDC:ResponseMode", OpenIdConnectResponseMode.Query);
+        options.RequireHttpsMetadata = !builder.Environment.IsDevelopment();
+        options.Scope.Clear();
+        builder.Configuration.GetValue("Authentication:OIDC:Scope", "openid profile email").Split(" ", StringSplitOptions.RemoveEmptyEntries).ToList().ForEach(s => options.Scope.Add(s));
+        options.SaveTokens = builder.Configuration.GetValue("Authentication:OIDC:SaveTokens",false);
+        options.Events = new OpenIdConnectEvents
+        {
             OnRemoteFailure = context =>
             {
                 var failure = context.Failure;